chore: update seatlayout with mode value and renamed the file
