+++ src/app/features/tripDetailsSlice.ts
