+++ src/app/store.ts
@@ -4,14 +4,14 @@ import { configureStore } from '@reduxjs/toolkit';
 import themeReducer from './features/themeSlice';
 import languageReducer from './features/languageSlice';
 import tripListingReducer from './features/tripListingSlice';
-import tripDetailsReducer from './features/tripdetailsSlice';
+import tripDetailsReducer from './features/tripDetailsSlice';
 
 export const store = configureStore({
     reducer: {
         theme: themeReducer,
         language: languageReducer,
         tripListing: tripListingReducer,
-        tripDetails: tripDetailsReducer
+        tripDetails: tripDetailsReducer,
     },
 });
