+++ src/components/SeatLayout/SeatLayout.tsx
@@ -1,9 +1,9 @@
 import { berthLayoutProducer, layoutConfig } from './seatConfig';
 import SeatLayoutWrapper from './SeatLayout.styled';
 import steeringWheel from '../../assets/tabler_steering-wheel.svg';
-import { ISeat, ISeatStatus } from '../../api/types/trip';
 import { useEffect, useState } from 'react';
 import Seat from './Seat/Seat';
+import { ISeat, ISeatStatus } from '../../types';
 
 const SeatLayout = ({
     layoutName,
@@ -48,41 +48,33 @@ const SeatLayout = ({
                             {row.map((seat: number, index: number) => {
                                 if (seat === 1) {
                                     const currentSeat = seats[seatIndex];
+                                    const seatStatus = (() => {
+                                        switch (currentSeat.status) {
+                                            case ISeatStatus.AVAILABLE:
+                                                if (
+                                                    selectedSeats.includes(
+                                                        currentSeat.seatNumber
+                                                    )
+                                                ) {
+                                                    return 'selected';
+                                                } else {
+                                                    return 'available';
+                                                }
+                                            case ISeatStatus.SELECTED:
+                                                return 'selected';
+                                            case ISeatStatus.BOOKED:
+                                                return 'unavailable';
+                                        }
+                                    })();
                                     seatIndex++;
-                                    return currentSeat.status.toString() ===
-                                        ISeatStatus.AVAILABLE.toString() ? (
-                                        selectedSeats.includes(
-                                            currentSeat.seatNumber
-                                        ) ? (
-                                            <Seat
-                                                seatNumber={
-                                                    currentSeat.seatNumber
-                                                }
-                                                key={`seat${rowIndex}-${index}`}
-                                                updateSelectedSeats={
-                                                    updateSelectedSeats
-                                                }
-                                                seatStatus={'selected'}
-                                                mode={mode}
-                                            />
-                                        ) : (
-                                            <Seat
-                                                seatNumber={
-                                                    currentSeat.seatNumber
-                                                }
-                                                key={`seat${rowIndex}-${index}`}
-                                                updateSelectedSeats={
-                                                    updateSelectedSeats
-                                                }
-                                                seatStatus={'available'}
-                                                mode={mode}
-                                            />
-                                        )
-                                    ) : (
+                                    return (
                                         <Seat
                                             seatNumber={currentSeat.seatNumber}
                                             key={`seat${rowIndex}-${index}`}
-                                            seatStatus={'unavailable'}
+                                            updateSelectedSeats={
+                                                updateSelectedSeats
+                                            }
+                                            seatStatus={seatStatus}
                                             mode={mode}
                                         />
                                     );
